+++ src/hooks/useRequest.tsx
@@ -1,5 +1,26 @@
-function useRequest() {
-  return <div>useRequest</div>;
-}
+// import { useState } from "react";
 
-export default useRequest;
+// function useRequest(getFunction: (i?: string) => any) {
+//   const [data, setData] = useState();
+//   const [isLoading, setIsLoading] = useState(false);
+//   const [error, setIsError] = useState<unknown>();
+
+//   const fetcher = async () => {
+//     try {
+//       setIsLoading(true);
+//       const result = getFunction();
+//       if (result) {
+//         setData(result);
+//         return data;
+//       }
+//     } catch (e: unknown) {
+//       setIsError(e);
+//     } finally {
+//       setIsLoading(false);
+//     }
+//   };
+
+//   return { data, isLoading, error, fetcher };
+// }
+
+// export default useRequest;
